requirements.txt
streamlit==1.41.0
google-generativeai==0.7.0
yfinance==0.2.32
pandas==2.2.0
